backend/requirements_new.txt
google-generativeai==0.8.5
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==21.1
nest-asyncio==1.6.0
psycopg2-binary==2.9.10
sqlalchemy>=2.0.25
pydantic==2.11.9
pydantic-settings==2.11.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-text-splitters==0.3.11
pypdf==6.1.1
faiss-cpu==1.12.0
pgvector==0.4.1
azure-identity>=1.15.0
twilio==9.8.3